requirements.txt
langchain
langchain-community
langchain-openai
chromadb
sentence-transformers
openai
tiktoken
python-dotenv
pypdf
pandas